requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
packaging==21.3
pluggy==1.0.0
protobuf==4.21.2
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
tomli==2.0.1
Werkzeug==2.1.2
